将带bottomBar的自定义一个BottomBarViewController 来处理，其他需要的继承该类

+++ OSChina/OSCUser.h
@@ -10,12 +10,12 @@
 
 @interface OSCUser : OSCBaseObject
 
-@property (nonatomic, assign, readonly) int64_t  userID;
-@property (nonatomic, strong, readonly) NSString *location;
-@property (nonatomic, strong, readonly) NSString *name;
-@property (nonatomic, assign, readonly) unsigned long     followersCount;
-@property (nonatomic, assign, readonly) unsigned long     fansCount;
-@property (nonatomic, assign ,readonly) long     score;
-@property (nonatomic, copy, readonly  ) NSURL    *portraitURL;
+@property (nonatomic, assign, readonly) int64_t    userID;
+@property (nonatomic, strong, readonly) NSString   *location;
+@property (nonatomic, strong, readonly) NSString   *name;
+@property (nonatomic, assign, readonly) NSUInteger followersCount;
+@property (nonatomic, assign, readonly) NSUInteger fansCount;
+@property (nonatomic, assign ,readonly) NSInteger  score;
+@property (nonatomic, copy, readonly  ) NSURL      *portraitURL;
 
 @end
